Make IGDB reconciliation more stable.
Switch to a concurrent model for reconciling many entries that has bound number of tasks instead of spawning unbounded tasks. The latter while query limited to 4 QPS it was exceeding the 8 active connection to the service and was violating the free tier quotas resulting to failures.

Additionally, recon now retrieves game screenshots.