Refactor to simplify LibraryViewModel.
This is now significantly simpler. It has two main input sources; the main user library and custom views that can show arbitrary group of games.

LibraryFilterModel becomes obsolete with this changes and needs cleanup. Same for RemoteLibraryModel.

RemoteLibraryModel was only used for collections at this point and will be replaced with a dedicated collections page.

The filter has been largely been obsolete with refinements widget controls and dedicated company pages. Collection views are temporarily not going to work until the dedicated collection page is created.